[refactor] controller 리팩토링
